Keep track of snapshot usage
Make sure that Snaphots in use by images are not being tagged for
deletion by housekeeper. Snapshots have not been accidentally
deleted previously, as the APi doesn't allow deletion of snapshots
in use. But this will make sure the snapshots are not even tagged
to start with (this also prevents these snapshots from being included
in deletion warnings).

Change-Id: I889289d87ade39b6a8179e72d6d48ba983314fba
Jira: None