Implement support for GCP
Add GCP support to Olga, which allows cleanup of resource in
Google Cloud. Support for the following is implemented:

	- Instances
	- Volumes/Disks
	- Images
	- Snapshots
	- Buckets (partially)

Tagging is not supported for buckets, as OWNER permissions are
required to change the buckets — this is too much to ask for.

Running against GCP is currently pretty useless, as there is no
file containing which projects to run against. @olle is working on
a new accounts file that contains both AWS and GCP accounts.

Change-Id: I62af92ba605d90c0f07814aaaa8689a1ad028b9a
Jira: YOLO-967