Don't tag resources if total cost < $100
Don't tag a users resources if the total cost of all the resources
is less than $100.

This is meant to decrease the number of emails sent to users.

This is not tested - not sure if there's a good way to test this
or if this can be seen as a low risk commit.

Change-Id: Ibad8605ee72c41f7a3aa62b5dc64dc8d4b95f63d
Jira: BENT-522